Move content module and include methods.
'prepare_mailing' gets all its info from the command line which in turn
relies on the dict in 'content' having cust_funcs as first class objects
(methods of the Membership class.)
Currently working for 'request_inductee_payment'.

+++ backup.sh
@@ -10,6 +10,11 @@ DEST=${BACKUP}/$STAMP
 LAST=`cat ${BACKUP}/last`
 echo $STAMP > ${BACKUP}/last
 
+if [-d ${DEST}]; then
+  echo Backup already done today. No backup until tomorrow.
+  exit 1
+fi
+
 mkdir $DEST
 cp -al ${BACKUP}/${LAST}/. ${BACKUP}/$DEST
 rsync -av --exclude='Py' --delete ../../Mshp/  $DEST
